Fix double-click for open, tweak wording
