+++ plugins/com.google.cloud.tools.eclipse.appengine.facets/src/com/google/cloud/tools/eclipse/appengine/facets/ui/navigator/AppEngineContentProvider.java
@@ -5,6 +5,12 @@ import com.google.cloud.tools.appengine.AppEngineDescriptor;
 import com.google.cloud.tools.appengine.api.AppEngineException;
 import com.google.cloud.tools.eclipse.appengine.facets.AppEngineStandardFacet;
 import com.google.cloud.tools.eclipse.appengine.facets.WebProjectUtil;
+import com.google.cloud.tools.eclipse.appengine.facets.ui.navigator.model.AppEngineCronDescriptor;
+import com.google.cloud.tools.eclipse.appengine.facets.ui.navigator.model.AppEngineDatastoreIndexes;
+import com.google.cloud.tools.eclipse.appengine.facets.ui.navigator.model.AppEngineDenialOfServiceDescriptor;
+import com.google.cloud.tools.eclipse.appengine.facets.ui.navigator.model.AppEngineDispatchDescriptor;
+import com.google.cloud.tools.eclipse.appengine.facets.ui.navigator.model.AppEngineTaskQueuesDescriptor;
+import com.google.cloud.tools.eclipse.appengine.facets.ui.navigator.model.AppEngineWebDescriptor;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
