+++ plugins/com.google.cloud.tools.eclipse.appengine.facets/src/com/google/cloud/tools/eclipse/appengine/facets/ui/navigator/model/AppEngineDenialOfServiceDescriptor.java
@@ -14,17 +14,16 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.eclipse.appengine.facets.ui.navigator;
+package com.google.cloud.tools.eclipse.appengine.facets.ui.navigator.model;
 
 import org.eclipse.core.resources.IFile;
 import org.eclipse.wst.common.project.facet.core.IFacetedProject;
 
 /**
- * Represents a {@code dispatch.xml} element.
+ * Represents a {@code dos.xml} element.
  */
-public class AppEngineDispatchDescriptor extends AppEngineResourceElement {
-  public AppEngineDispatchDescriptor(IFacetedProject project, IFile file) {
+public class AppEngineDenialOfServiceDescriptor extends AppEngineResourceElement {
+  public AppEngineDenialOfServiceDescriptor(IFacetedProject project, IFile file) {
     super(project, file);
   }
-
 }
