+++ plugins/com.google.cloud.tools.eclipse.appengine.facets/src/com/google/cloud/tools/eclipse/appengine/facets/ui/navigator/model/AppEngineTaskQueuesDescriptor.java
@@ -14,26 +14,16 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.eclipse.appengine.facets.ui.navigator;
+package com.google.cloud.tools.eclipse.appengine.facets.ui.navigator.model;
 
-import com.google.cloud.tools.appengine.AppEngineDescriptor;
-import com.google.common.base.Preconditions;
 import org.eclipse.core.resources.IFile;
 import org.eclipse.wst.common.project.facet.core.IFacetedProject;
 
 /**
- * A model representation of the {@code appengine-web.xml}.
+ * Represents an {@code queue.xml} element.
  */
-public class AppEngineWebDescriptor extends AppEngineResourceElement {
-  private final AppEngineDescriptor descriptor;
-
-  public AppEngineWebDescriptor(IFacetedProject project, IFile file,
-      AppEngineDescriptor descriptor) {
+public class AppEngineTaskQueuesDescriptor extends AppEngineResourceElement {
+  public AppEngineTaskQueuesDescriptor(IFacetedProject project, IFile file) {
     super(project, file);
-    this.descriptor = Preconditions.checkNotNull(descriptor);
-  }
-
-  public AppEngineDescriptor getDescriptor() {
-    return descriptor;
   }
 }
